# docs/persona-ai-mvp.openapi.yaml
openapi: 3.0.3
info:
  title: TalkTo Persona AI MVP API
  version: 0.1.0
  description: |
    Backend MVP API design for Persona AI.

    v1.0 source priority: TalkTo_개발자용_기능명세서_v1.0.md overrides older Notion rows,
    pasted planning notes, and the legacy MVP baseline where they conflict.

    Current implemented surface: legacy Persona Chat MVP plus Archive-first
    subjects, purpose consents, question cards/interactions, recording upload,
    recording completion, recording list/detail, original recording playback,
    payment products/orders/webhooks/entitlements, and queued paid Archive
    analysis jobs, analysis job status APIs, audited admin retry hooks, and
    admin worker transition endpoints for paid analysis jobs, and paid Memories
    status/search plus segment playback/feedback over indexed analysis segments,
    in-app analysis status notifications, and R&D redacted export opt-in/preview
    guardrails, plus Voice Persona application core routes.

    Guardrails: free Archive upload/completion must not create Preview, STT,
    LLM summary, embedding, memory segment, or analysis job work. Preview/sample
    analysis is deferred/disabled by v1.0. POST
    /api/v1/recordings/{recordingId}/preview-analysis exists only as a disabled
    compatibility endpoint returning 409 feature_deferred and must not enqueue
    Preview, free AI, or analysis job work.

    Not yet implemented: automated provider execution for analysis workers,
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/voice-persona/applications:
    post:
      tags: [Voice Persona]
      summary: Create a Voice Persona build application
      description: Requires subject ownership, required Voice Persona consents, and an active Voice Persona entitlement. Creates a locked build application and locked build job; it does not enable runtime chat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoicePersonaApplicationRequest'
      responses:
        '201':
          description: Voice Persona application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoicePersonaApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/voice-persona/applications/{applicationId}/documents/upload-intent:
    post:
      tags: [Voice Persona]
      summary: Create a Voice Persona evidence document upload intent
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoicePersonaDocumentIntentRequest'
      responses:
        '201':
          description: Document upload intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoicePersonaDocumentIntentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/voice-persona/applications/{applicationId}/intake:
    put:
      tags: [Voice Persona]
      summary: Save Voice Persona intake draft sections
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPersonaIntakeRequest'
      responses:
        '200':
          description: Intake draft saved
  /api/v1/voice-persona/applications/{applicationId}/intake/submit:
    post:
      tags: [Voice Persona]
      summary: Submit complete 8-section Persona intake
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      responses:
        '200':
          description: Intake submitted
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/voice-persona/applications/{applicationId}/voice-samples:
    post:
      tags: [Voice Persona]
      summary: Submit target voice sample for ops review
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetVoiceSampleRequest'
      responses:
        '201':
          description: Target voice sample submitted
  /api/v1/voice-persona/applications/{applicationId}/build-status:
    get:
      tags: [Voice Persona]
      summary: Get Voice Persona build status lock reasons
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      responses:
        '200':
          description: Voice Persona build status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonaBuildStatusResponse'
  /api/v1/voice-persona/applications/{applicationId}/family-review:
    get:
      tags: [Voice Persona]
      summary: Get family-safe Persona Bible review view
      description: Returns only approved Persona Bible summary/safety notes for owner review. Raw evidence, recordings, transcripts, and provider payloads are not returned.
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      responses:
        '200':
          description: Family review view
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/voice-persona/applications/{applicationId}/family-review/approve:
    post:
      tags: [Voice Persona]
      summary: Approve family review and enable runtime config when gates pass
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      responses:
        '200':
          description: Review approved and runtime config enabled if all gates pass
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/voice-persona/applications/{applicationId}/family-review/request-changes:
    post:
      tags: [Voice Persona]
      summary: Request Persona Bible changes before runtime enablement
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPersonaBibleChangesRequest'
      responses:
        '200':
          description: Change request persisted
  /api/v1/admin/voice-persona/documents/{documentId}/review:
    patch:
      tags: [Admin, Voice Persona]
      summary: Review Voice Persona evidence document
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewVoicePersonaAssetRequest'
      responses:
        '200':
          description: Document reviewed
  /api/v1/admin/voice-persona/voice-samples/{sampleId}/review:
    patch:
      tags: [Admin, Voice Persona]
      summary: Review target voice sample
      parameters:
        - in: path
          name: sampleId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewVoicePersonaAssetRequest'
      responses:
        '200':
          description: Voice sample reviewed
  /api/v1/admin/voice-persona/applications/{applicationId}/build-status:
    patch:
      tags: [Admin, Voice Persona]
      summary: Manually update Voice Persona build status
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonaBuildStatusRequest'
      responses:
        '200':
          description: Build status updated
  /api/v1/admin/voice-persona/applications/{applicationId}/persona-bible:
    put:
      tags: [Admin, Voice Persona]
      summary: Create or update Persona Bible draft
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPersonaBibleRequest'
      responses:
        '200':
          description: Persona Bible saved
  /api/v1/admin/voice-persona/persona-bibles/{bibleId}/review:
    patch:
      tags: [Admin, Voice Persona]
      summary: Review Persona Bible
      parameters:
        - in: path
          name: bibleId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewVoicePersonaAssetRequest'
      responses:
        '200':
          description: Persona Bible reviewed
  /api/v1/admin/voice-persona/applications/{applicationId}/provider-assets:
    post:
      tags: [Admin, Voice Persona]
      summary: Register manual voice provider asset
      parameters:
        - $ref: '#/components/parameters/VoicePersonaApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterVoiceProviderAssetRequest'
      responses:
        '200':
          description: Provider asset registered
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or token is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
...
    VoicePersonaApplicationId:
      in: path
      name: applicationId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: CONFLICT
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
    CreateVoicePersonaApplicationRequest:
      type: object
      required: [subjectId]
      properties:
        subjectId:
          type: string
          format: uuid
    VoicePersonaApplicationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/VoicePersonaApplication'
    VoicePersonaApplication:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        entitlementId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [draft, locked_until_requirements, ready_for_admin_review, build_pending, ready]
        documentsStatus:
          type: string
          enum: [pending_review, approved, rejected]
        intakeStatus:
          type: string
          enum: [draft, submitted]
        voiceSampleStatus:
          type: string
          enum: [pending_review, approved, rejected]
        buildStatus:
          type: string
          enum: [locked_until_requirements, pending_admin_review, pending_manual_provider_asset, ready, failed]
    CreateVoicePersonaDocumentIntentRequest:
      type: object
      required: [filename, mimeType, fileSizeBytes]
      properties:
        filename:
          type: string
          maxLength: 255
        mimeType:
          type: string
          maxLength: 120
        fileSizeBytes:
          type: string
    VoicePersonaDocumentIntentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            document:
              type: object
            uploadUrl:
              type: string
            method:
              type: string
              enum: [PUT]
    UpsertPersonaIntakeRequest:
      type: object
      required: [sections]
      properties:
        sections:
          type: array
          minItems: 1
          items:
            type: object
            required: [sectionKey, answers]
            properties:
              sectionKey:
                type: string
                maxLength: 80
              answers:
                type: object
                additionalProperties: true
    CreateTargetVoiceSampleRequest:
      type: object
      properties:
        recordingId:
          type: string
          format: uuid
        storageKey:
          type: string
          maxLength: 500
    PersonaBuildStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            applicationId:
              type: string
              format: uuid
            buildStatus:
              type: string
            documentsStatus:
              type: string
            intakeStatus:
              type: string
            voiceSampleStatus:
              type: string
            lockedReasons:
              type: array
              items:
                type: string
    RequestPersonaBibleChangesRequest:
      type: object
      properties:
        notes:
          type: string
          maxLength: 1000
    ReviewVoicePersonaAssetRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending_review, approved, rejected]
        notes:
          type: string
          maxLength: 500
    UpdatePersonaBuildStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [locked_until_requirements, pending_admin_review, pending_manual_provider_asset, ready, failed]
    UpsertPersonaBibleRequest:
      type: object
      required: [contentSummary]
      properties:
        contentSummary:
          type: string
          maxLength: 10000
        safetyNotes:
          type: string
          maxLength: 2000
    RegisterVoiceProviderAssetRequest:
      type: object
      required: [providerName, externalAssetId]
      properties:
        providerName:
          type: string
          maxLength: 120
        externalAssetId:
          type: string
          maxLength: 160
        status:
          type: string
          enum: [pending_manual_registration, registered, failed, deletion_requested, deleted]
        notes:
          type: string
          maxLength: 1000
...